requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
